+++ changes.md
@@ -6,6 +6,7 @@
 * Adopt pip-tools to manage requirements.txt files
 * Refactor statistics reporter to support dynamically added metrics
 * Add latency and per-PV graphite metrics
+* add support for streaming PV units (`un00` schema)
 
 ## v2.1.0